requirements.txt
asgiref==3.6.0
dj-database-url==2.0.0
Django==3.2.19
django-ninja==0.21.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
pycryptodome==3.18.0
pydantic==1.10.7
sqlparse==0.4.4
typing_extensions==4.5.0
whitenoise==6.4.0
